Rename package and add license header
