+++ elasticsearch-jest/src/main/java/org/apache/metamodel/jest/elasticsearch/JestElasticSearchInsertBuilder.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.metamodel.jest.elasticsearch;
+package org.apache.metamodel.jest.elasticsearch;
 
 import io.searchbox.core.DocumentResult;
 import io.searchbox.core.Index;
