+++ elasticsearch-jest/src/main/java/org/apache/metamodel/jest/elasticsearch/JestElasticSearchMetaData.java
@@ -16,19 +16,35 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.metamodel.jest.elasticsearch;
+package org.apache.metamodel.jest.elasticsearch;
 
-import java.util.Date;
+import org.apache.metamodel.schema.ColumnType;
 
-import junit.framework.TestCase;
+/**
+ * MetaData representation of an ElasticSearch index type.
+ *
+ * We will map the elasticsearch fields to columns and their
+ * types to {@link ColumnType}s.
+ */
+public class JestElasticSearchMetaData {
+    
+    private final String[] columnNames;
+    private final ColumnType[] columnTypes;
 
-public class JestElasticSearchDateConverterTest extends TestCase {
+    /**
+     * Constructs a {@link JestElasticSearchMetaData}.
+     *
+     */
+    public JestElasticSearchMetaData(String[] names, ColumnType[] types) {
+        this.columnNames = names;
+        this.columnTypes = types;
+    }
 
-    public void testConvertDateOptionalTime() throws Exception {
-        String dateToConvert = "2013-01-04T15:55:51.217+01:00";
-        Date date = JestElasticSearchDateConverter.tryToConvert(dateToConvert);
+    public String[] getColumnNames() {
+        return columnNames;
+    }
 
-        assertNotNull(date);
-        assertTrue(date.toString().startsWith("Fri Jan 04"));
+    public ColumnType[] getColumnTypes() {
+        return columnTypes;
     }
 }
